Move SenseProcessorUtils to suripu.core so it can be used from Firehose
